Add generic tests
